JX-5: Decimal functionality

+++ src/main/java/oisin/connery/structures/NumberAndIndexes.java
@@ -3,10 +3,12 @@ package oisin.connery.structures;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Builder
 public class NumberAndIndexes { // make valueAndIndexes and make value generic?
-    private int number;
+    private BigDecimal number; // todo: make generic number?
     private Integer startingIndex;
     private Integer endingIndex;
 }
